% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_all}
\alias{compare_all}
\title{Performs whole analysis for given mosquito and human datasets}
\usage{
compare_all(mos, hum, threshold, sixteen)
}
\arguments{
\item{mos}{All mosquito marker sizes}

\item{hum}{All human marker sizes}

\item{threshold}{Maximum allowable distance for a match}

\item{sixteen}{Boolean denoting 10 or 17 markers used}
}
\description{
Performs whole analysis for given mosquito and human datasets
}
