% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_marker}
\alias{compare_marker}
\title{Compares allele sizes at a single loci}
\usage{
compare_marker(hm, mm)
}
\arguments{
\item{hm}{Human marker}

\item{mm}{Mosquito marker}
}
\description{
Compares allele sizes at a single loci
}
