% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_mosquito}
\alias{compare_mosquito}
\title{Compares all 10 markers in a mosquito to all 10 markers in a human}
\usage{
compare_mosquito(human, mosquito)
}
\arguments{
\item{human}{Table of marker sizes for human}

\item{mosquito}{Table of marker sizes for mosquito}
}
\description{
Compares all 10 markers in a mosquito to all 10 markers in a human
}
