% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_mosquito_16}
\alias{compare_mosquito_16}
\title{Compares all 17 markers in a mosquito to all 17 markers in a human}
\usage{
compare_mosquito_16(human, mosquito)
}
\arguments{
\item{human}{Table of human marker sizes}

\item{mosquito}{Table of mosquito marker sizes}
}
\description{
Compares all 17 markers in a mosquito to all 17 markers in a human
}
