% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{human_matches}
\alias{human_matches}
\title{Compares a mosquito to all humans that live in the house within which it was caught}
\usage{
human_matches(mosquito, humans, threshold, sixteen)
}
\arguments{
\item{mosquito}{Mosquito marker data}

\item{humans}{All human marker data}

\item{threshold}{Distance below which a mosquito and human are declared a match}

\item{sixteen}{boolean value denoting whether 10 or 17 markers are being used}
}
\description{
Compares a mosquito to all humans that live in the house within which it was caught
}
