% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_humans}
\alias{read_humans}
\title{Reads in csv file data for humans}
\usage{
read_humans(sixteen)
}
\arguments{
\item{sixteen}{boolean denoting whether use wants 10 or 17 marker data}
}
\description{
Reads in csv file data for humans
}
