% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_mosquitoes}
\alias{read_mosquitoes}
\title{Reads in csv file of mosquito data}
\usage{
read_mosquitoes(fname)
}
\arguments{
\item{fname}{string input of filename}
}
\description{
Reads in csv file of mosquito data
}
