% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_humans}
\alias{read_humans}
\title{Reads in csv file data for humans}
\usage{
read_humans(fname, agefile)
}
\arguments{
\item{fname}{string input filename}

\item{agefile}{string input file containing corresponding human ages}
}
\description{
Reads in csv file data for humans
}
