% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_mosquitoes}
\alias{read_mosquitoes}
\title{Reads in csv file of mosquito data}
\usage{
read_mosquitoes(sixteen)
}
\arguments{
\item{sixteen}{boolean specifying whether user wants 10 or 17 marker dataset}
}
\description{
Reads in csv file of mosquito data
}
